refactor: :recycle: Refactory code
