+++ src/raspberry/database/operaciones.py
