+++ src/raspberry/deteccion/haarcascade_frontalface_alt.xml
